+++ composer/algorithms/__init__.py
@@ -62,6 +62,7 @@ from composer.algorithms.hparams import CutMixHparams as CutMixHparams
 from composer.algorithms.hparams import CutOutHparams as CutOutHparams
 from composer.algorithms.hparams import LabelSmoothingHparams as LabelSmoothingHparams
 from composer.algorithms.hparams import MixUpHparams as MixUpHparams
+from composer.algorithms.hparams import RandAugmentHparams as RandAugmentHparams
 from composer.algorithms.label_smoothing import LabelSmoothing as LabelSmoothing
 from composer.algorithms.layer_freezing import LayerFreezing as LayerFreezing
 from composer.algorithms.layer_freezing import LayerFreezingHparams as LayerFreezingHparams
@@ -71,7 +72,6 @@ from composer.algorithms.no_op_model import NoOpModelHparams as NoOpModelHparams
 from composer.algorithms.progressive_resizing import ProgressiveResizing as ProgressiveResizing
 from composer.algorithms.progressive_resizing import ProgressiveResizingHparams as ProgressiveResizingHparams
 from composer.algorithms.randaugment import RandAugment as RandAugment
-from composer.algorithms.randaugment import RandAugmentHparams as RandAugmentHparams
 from composer.algorithms.sam import SAM as SAM
 from composer.algorithms.sam import SAMHparams as SAMHparams
 from composer.algorithms.scale_schedule import ScaleSchedule as ScaleSchedule
