+++ composer/algorithms/randaugment/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.randaugment.randaugment import RandAugment as RandAugment
-from composer.algorithms.randaugment.randaugment import RandAugmentHparams as RandAugmentHparams
 from composer.algorithms.randaugment.randaugment import RandAugmentTransform as RandAugmentTransform
 from composer.algorithms.randaugment.randaugment import randaugment_image as randaugment_image
